move version package to be more general
