+++ version/strings.go
